backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.6
alembic==1.11.1
pgvector==0.2.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic==2.5.3
pydantic-settings==2.1.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.2.0
pytesseract==0.3.10
openai==1.10.0
python-dotenv==1.0.0
numpy==1.26.3
scikit-learn==1.4.0
scipy==1.11.4
aiofiles==23.2.1
email-validator==2.1.0.post1
httpx>=0.25,<0.28

